+++ flask_cuttlepool.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 """
-Flask extension for CuttlePool. This is based in large part on the SQLite3
+Flask extension for CuttlePool. This extension is inspired by the SQLite3
 example given in the `Flask Extension Development
-<http://flask.pocoo.org/docs/0.12/extensiondev>`_ documentation.
+<http://flask.pocoo.org/docs/0.12/extensiondev>`_ documentation and
+`Flask-SQLAlchemy <http://flask-sqlalchemy.pocoo.org>`_.
 
 :license: BSD 3-clause, see LICENSE for details.
 """
@@ -10,7 +11,10 @@ example given in the `Flask Extension Development
 __version__ = '0.2.0-dev'
 
 
+from threading import RLock
+
 from cuttlepool import CuttlePool, CuttlePoolError, PoolConnection
+from flask import current_app
 
 try:
     from cuttlepool import _CAPACITY, _OVERFLOW, _TIMEOUT
@@ -45,12 +49,14 @@ class FlaskCuttlePool(object):
 
     def __init__(self, connect, capacity=_CAPACITY, overflow=_OVERFLOW,
                  timeout=_TIMEOUT, app=None, **kwargs):
-        self._app_set = False
-        self._connect = connect,
-        self._capacity = capacity
-        self._timeout = timeout
+        self._connect = connect
         self._app = app
-        self._connection_arguments = kwargs
+        self._cuttlepool_kwargs = kwargs
+        self._cuttlepool_kwargs.update(capacity=capacity,
+                                       overflow=overflow,
+                                       timeout=timeout)
+
+        self._lock = RLock()    # Necessary for multithreaded apps.
 
         if app is not None:
             self.init_app(app)
@@ -75,13 +81,6 @@ class FlaskCuttlePool(object):
         # pool will connect to steakhouse instead.
         pool.init_app(app)
         """
-        prefix = 'CUTTLEPOOL_'
-        self._app_set = True
-        self._connection_arguments.update(
-            **{k[len(prefix):].lower(): v
-               for k, v in app.config.items()
-               if k.startswith(prefix)})
-
         # Use the newstyle teardown_appcontext if it's available,
         # otherwise fall back to the request context.
         if hasattr(app, 'teardown_appcontext'):
@@ -89,44 +88,58 @@ class FlaskCuttlePool(object):
         else:
             app.teardown_request(self.teardown)
 
+    def _get_app(self):
+        """
+        Looks up the current application or the default passed to
+        ``__init__()``
+        """
+        if current_app:
+            return current_app._get_current_object()
+
+        if self._app is not None:
+            return self._app
+
+        raise RuntimeError('No application found.')
+
+    def _get_pool(self):
+        """Gets the pool on the current application."""
+        app = self._get_app()
+
+        with self._lock:
+            if not hasattr(app, 'extensions'):
+                app.extensions = {}
+
+            if 'cuttlepool' not in app.extensions:
+                app.extensions['cuttlepool'] = self._make_pool()
+
+            return app.extensions['cuttlepool']
+
+    def _make_pool(self):
+        """Make a CuttlePool instance."""
+        prefix = 'CUTTLEPOOL_'
+        kwargs = self._cuttlepool_kwargs.copy()
+
+        app = self._get_app()
+
+        kwargs.update(
+            **{k[len(prefix):].lower(): v
+               for k, v in app.config.items()
+               if k.startswith(prefix)})
+
+        return CuttlePool(self._connect, **kwargs)
+
+    def cursor(self):
+        """Gets a cursor from the connection on the appplication context."""
+        return self.connection.cursor()
+
     def get_connection(self):
         """
-        Gets a ``PoolConnection`` object. Saves the connection on the
-        application context for subsequent calls.
-
-        :raise AppUninitializedError: If trying to get a connection before
-                                      calling ``init_app()``.
+        Gets a ``PoolConnection`` object. The caller of this method is
+        responsible for calling the ``close()`` method on the
+        ``PoolConnection`` object.
         """
-        if not self._app_set:
-            raise AppUninitializedError('init_app() with an app object before using the pool')
-
-        ctx = stack.top
-
-        if ctx is not None:
-            if (not hasattr(ctx, 'cuttlepool_connection') or
-                    ctx.cuttlepool_connection._connection is None):
-                ctx.cuttlepool_connection = super(FlaskCuttlePool, self).get_connection()
-
-            else:
-                if not self.ping(ctx.cuttlepool_connection):
-                    ctx.cuttlepool_connection.close()
-                    del ctx.cuttlepool_connection
-                    ctx.cuttlepool_connection = self.get_connection()
-
-            return ctx.cuttlepool_connection
-
-    def get_fresh_connection(self):
-        """
-        Gets a ``PoolConnection`` object directly from the pool. Does not try
-        to retrieve from the application context, nor does it store the
-        connection on the application context.
-        :raise AppUninitializedError: If trying to get a connection before
-                                      calling ``init_app()``.
-        """
-        if not self._app_set:
-            raise AppUninitializedError('init_app() with an app object before using the pool')
-
-        return super(FlaskCuttlePool, self).get_connection()
+        pool = self._get_pool()
+        return pool.get_connection()
 
     def teardown(self, exception):
         """
@@ -138,12 +151,24 @@ class FlaskCuttlePool(object):
         if hasattr(ctx, 'cuttlepool_connection'):
             ctx.cuttlepool_connection.close()
 
+    @property
+    def connection(self):
+        """
+        Gets a ``PoolConnection`` object. Saves the connection on the
+        application context for subsequent gets.
 
-class FlaskCuttlePoolError(CuttlePoolError):
+        If there is no application context, returns ``None``.
+        """
+        ctx = stack.top
 
-    """Base class for exceptions in this module."""
+        if ctx is not None:
+            if (not hasattr(ctx, 'cuttlepool_connection') or
+                    ctx.cuttlepool_connection._connection is None):
+                ctx.cuttlepool_connection = self.get_connection()
 
+            else:
+                if not self._get_pool().ping(ctx.cuttlepool_connection):
+                    ctx.cuttlepool_connection.close()
+                    ctx.cuttlepool_connection = self.connection
 
-class AppUninitializedError(FlaskCuttlePoolError):
-
-    """Exception raised when the pool is used before calling ``init_app()``."""
+            return ctx.cuttlepool_connection