+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 ### Added
-- Import `PoolConnection` for simple, uniform import from `flask_cuttlepool` by
-  ([#4](https://github.com/smitchell556/flask-cuttlepool/pull/4)).
+- Import `PoolConnection` for simple, uniform import from `flask_cuttlepool`.
+  ([#4](https://github.com/smitchell556/flask-cuttlepool/pull/4))
+- Support for multiple Flask applications. Connection pools are created per
+  application.
 ### Changed
 - Upgrade minimum version of `cuttlepool` to 0.6.0.
 - Get all configuration options starting with `CUTTLEPOOL_` from `app.config`