+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
-## [Unreleased]
+## [0.2.0] - 2018-01-30
 ### Added
 - Import `PoolConnection` for simple, uniform import from `flask_cuttlepool`.
   ([#4](https://github.com/smitchell556/flask-cuttlepool/pull/4))
@@ -23,4 +23,4 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ### Added
 - Initial Flask-CuttlePool code.
 
-[Unreleased]: https://github.com/smitchell556/flask-cuttlepool/compare/v0.1.0...HEAD
+[0.2.0]: https://github.com/smitchell556/flask-cuttlepool/compare/v0.1.0...0.2.0